clean-repo/backend/requirements.txt
Flask
Flask-Cors
Flask-SQLAlchemy
SQLAlchemy
flask-jwt-extended
python-dotenv
psycopg2-binary
Werkzeug
reportlab
Pillow
boto3
botocore
marshmallow
marshmallow-sqlalchemy
charset-normalizer
reportlab
gunicorn
google-generativeai
Authlib